+++ docs/changelog.md
@@ -15,6 +15,7 @@ category: Administration
 * chore: improve Github action for continuous integration
 * chore: update code of conduct to version 3
 * feat: ensure compatibility with Jira 11.2.0
+* refactor: use Java 21 features for session mapper
 
 ### [25.10.0] - 2025-10-03